requirements.txt
numpy
Pillow
opencv-python
keyboard
pynput
screeninfo

# Minimum Python version
python_version >= "3.6"